adding insurer types: context | provider | i18n
