+++ packages/core/src/types/common/i18n/index.ts
@@ -4,3 +4,4 @@ export * from './actuary';
 export * from './corporateGroup';
 export * from './independentTrustee';
 export * from './inHouse';
+export * from './insurer';
