+++ packages/core/src/types/common/i18n/insurer.ts
@@ -1,0 +1,59 @@
+export type InsurerI18nProps = {
+	preview: {
+		buttons: {
+			one: string;
+			two: string;
+			three: string;
+			four: string;
+			five: string;
+		};
+		checkboxLabel: string;
+	};
+	reference: {
+		title: string;
+		subtitle: string;
+		fields: {
+			insurerCompanyReference: {
+				label: string;
+				errorIfEmpty: string;
+				errorIfTooShort: string;
+				errorIfTooLong: string;
+			};
+		};
+	};
+	remove: {
+		confirm: {
+			breadcrumbs: {
+				link1: string;
+				link2: string;
+			};
+			title: string;
+			dialog: {
+				message1: string;
+				message2: string;
+			};
+			buttons: {
+				remove: string;
+				cancel: string;
+			};
+		};
+		date: {
+			title: string;
+			fields: {
+				confirm: {
+					label: string;
+				};
+				date: {
+					label: string;
+					hint: string;
+					error: string;
+				};
+			};
+			errors: {
+				formIncomplete: string;
+				dateAddedBeforeEffectiveDate: string;
+				dateAddedInTheFuture: string;
+			};
+		};
+	};
+};
