+++ packages/core/src/types/common/providers/index.ts
@@ -4,3 +4,4 @@ export * from './actuary';
 export * from './corporateGroup';
 export * from './independentTrustee';
 export * from './inHouse';
+export * from './insurer';
