+++ packages/layout/src/components/cards/insurer/context.tsx
@@ -1,16 +1,9 @@
-import React, { createContext, useContext, ReactElement } from 'react';
+import React, { createContext, useContext } from 'react';
 import { useMachine } from '@xstate/react';
-import insurerMachine, { InsurerContext as EC } from './insurerMachine';
-import { State, EventData } from 'xstate';
-import { i18n as i18nDefaults, InsurerI18nProps } from './i18n';
+import insurerMachine from './insurerMachine';
+import { i18n as i18nDefaults } from './i18n';
 import { useI18n } from '../hooks/use-i18n';
-import {
-	CardDefaultProps,
-	CardContactDetails,
-	CardAddress,
-	CardProviderProps,
-	RecursivePartial,
-} from '../common/interfaces';
+import { InsurerContextProps, InsurerProviderProps } from '@tpr/core';
 
 export const InsurerContext = createContext<InsurerContextProps>({
 	current: {},
@@ -20,30 +13,6 @@ export const InsurerContext = createContext<InsurerContextProps>({
 	onSaveRef: Promise.resolve,
 	i18n: i18nDefaults,
 });
-
-type RenderProps = (_props: InsurerContextProps) => ReactElement;
-
-export interface InsurerContextProps
-	extends Omit<InsurerProviderProps, 'insurer'> {
-	send: (event: any, payload?: EventData) => Partial<State<EC, any, any, any>>;
-	current: Partial<State<EC, any, any, any>>;
-}
-
-export interface Insurer extends CardDefaultProps, CardContactDetails {
-	organisationReference: number;
-	organisationName: string;
-	insurerCompanyReference: string;
-	address: Partial<CardAddress>;
-}
-
-export interface InsurerProviderProps extends CardProviderProps {
-	onSaveRef?: (...args: any[]) => Promise<any>;
-	/** insurer props from the API */
-	insurer: Partial<Insurer>;
-	children?: RenderProps | ReactElement;
-	/** overwrite any text that you need */
-	i18n?: RecursivePartial<InsurerI18nProps>;
-}
 
 export const InsurerProvider = ({
 	complete,
